# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CRM Proto API",
    "version": "1.0.0",
    "description": "# CRM Proto API\n\nREST API для CRM системы.\n\n## Аутентификация\n\nAPI использует cookie-based аутентификацию через NextAuth.js.\nДля доступа к защищённым эндпоинтам необходимо авторизоваться через `/login`.\n\n## Ошибки\n\nВсе ошибки возвращаются в формате:\n\n```json\n{\n  \"error\": \"Описание ошибки\"\n}\n```\n\nСтандартные коды ответов:\n- `200` - Успешный запрос\n- `201` - Ресурс создан\n- `400` - Неверный запрос (ошибка валидации)\n- `401` - Не авторизован\n- `403` - Доступ запрещён\n- `404` - Ресурс не найден\n- `500` - Внутренняя ошибка сервера"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "paths": {
...
    "/api/users/search": {
      "post": {
        "summary": "Поиск пользователей",
        "description": "Возвращает список пользователей с пагинацией. Фильтры передаются в теле запроса. Доступно только администраторам.",
        "tags": [
          "Пользователи"
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "Уникальный идентификатор"
                          },
                          "email": {
                            "type": "string",
                            "description": "Email"
                          },
                          "name": {
                            "type": "string",
                            "description": "Имя"
                          },
                          "image": {
                            "type": "string"
                          },
                          "roles": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "enum": [
                                "admin",
                                "manager",
                                "user"
                              ],
                              "description": "Роль пользователя: admin - администратор, manager - менеджер, user - пользователь"
                            },
                            "description": "Список ролей"
                          },
                          "isActive": {
                            "type": "boolean",
                            "description": "Активен ли пользователь"
                          },
                          "lastLoginAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "description": "Дата создания"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time",
                            "description": "Дата обновления"
                          }
                        },
                        "required": [
                          "id",
                          "email",
                          "name",
                          "roles",
                          "isActive",
                          "createdAt",
                          "updatedAt"
                        ],
                        "description": "Пользователь"
                      },
                      "description": "Список пользователей"
                    },
                    "total": {
                      "type": "number",
                      "description": "Общее количество"
                    },
                    "page": {
                      "type": "number",
                      "description": "Текущая страница"
                    },
                    "limit": {
                      "type": "number",
                      "description": "Размер страницы"
                    }
                  },
                  "required": [
                    "users",
                    "total",
                    "page",
                    "limit"
                  ],
                  "description": "Список пользователей с пагинацией"
                },
                "example": {
                  "data": [
                    {
                      "id": "clx1234567890abcdef",
                      "email": "admin@example.com",
                      "name": "Администратор Системы",
                      "role": "admin",
                      "isActive": true,
                      "createdAt": "2024-01-10T08:00:00.000Z",
                      "updatedAt": "2024-01-10T08:00:00.000Z"
                    }
                  ],
                  "total": 1,
                  "page": 1,
                  "limit": 20,
                  "totalPages": 1
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "search": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "manager",
                      "user"
                    ],
                    "description": "Роль пользователя: admin - администратор, manager - менеджер, user - пользователь"
                  },
                  "isActive": {
                    "type": "boolean"
                  },
                  "page": {
                    "type": "integer",
                    "minimum": 0,
                    "default": 1
                  },
                  "limit": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 100,
                    "default": 20
                  }
                },
                "description": "Фильтры для списка пользователей"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}